Move RejectionExponential into distributions.testing
