+++ tests/distributions/test_rejector.py
@@ -5,7 +5,7 @@ import torch
 from torch.autograd import Variable, grad
 
 from pyro.distributions import Exponential, Gamma
-from pyro.distributions.rejector import Rejector
+from pyro.distributions.testing.rejection_exponential import RejectionExponential
 from pyro.distributions.testing.rejection_gamma import RejectionStandardGamma
 from tests.common import assert_equal
 
@@ -19,36 +19,19 @@ def compute_elbo_grad(model, guide, variables):
     return grad(surrogate_elbo.sum(), variables, create_graph=True)
 
 
-@pytest.mark.parametrize('alpha', [1.0, 2.0, 5.0])
-def test_standard_gamma_elbo(alpha):
+@pytest.mark.parametrize('rate', [0.5, 1.0, 2.0])
+@pytest.mark.parametrize('factor', [0.25, 0.5, 1.0])
+def test_rejector(rate, factor):
     num_samples = 100000
-    alphas = Variable(torch.Tensor([alpha]).expand(num_samples, 1), requires_grad=True)
-    betas = Variable(torch.ones(num_samples, 1))
-
-    model = Gamma(Variable(torch.ones(num_samples, 1)), betas)
-    guide1 = Gamma(alphas, betas)
-    guide2 = RejectionStandardGamma(alphas)  # implemented using Rejector
-
-    grads = []
-    for guide in [guide1, guide2]:
-        grads.append(compute_elbo_grad(model, guide, [alphas])[0].data)
-    expected, actual = grads
-    assert_equal(actual.mean(), expected.mean(), prec=0.01, msg='bad grad for alpha')
-
-
-class RejectionExponential(Rejector):
-    def __init__(self, rate, factor):
-        assert (factor <= 1).all()
-        self.rate = rate
-        self.factor = factor
-        propose = Exponential(self.factor * self.rate)
-        log_scale = self.factor.log()
-        super(RejectionExponential, self).__init__(propose, self.log_prob_accept, log_scale)
-
-    def log_prob_accept(self, x):
-        result = (self.factor - 1) * self.rate * x
-        assert result.max() <= 0, result.max()
-        return result
+    rates = Variable(torch.Tensor([rate]).expand(num_samples, 1), requires_grad=True)
+    factors = Variable(torch.Tensor([factor]).expand(num_samples, 1), requires_grad=True)
+    dist1 = Exponential(rates)
+    dist2 = RejectionExponential(rates, factors)
+    x1 = dist1.sample()
+    x2 = dist2.sample()
+    assert_equal(x1.mean(), x2.mean(), prec=0.02, msg='bug in .sample()')
+    assert_equal(x1.std(), x2.std(), prec=0.02, msg='bug in .sample()')
+    assert_equal(dist1.batch_log_pdf(x1), dist2.batch_log_pdf(x1), msg='bug in .batch_log_pdf()')
 
 
 @pytest.mark.xfail
@@ -74,16 +57,18 @@ def test_exponential_elbo(rate, factor):
     assert_equal(actual.mean(), 0.0, prec=0.05, msg='bad grad for factor')
 
 
-@pytest.mark.parametrize('rate', [0.5, 1.0, 2.0])
-@pytest.mark.parametrize('factor', [0.25, 0.5, 1.0])
-def test_rejector(rate, factor):
+@pytest.mark.parametrize('alpha', [1.0, 2.0, 5.0])
+def test_standard_gamma_elbo(alpha):
     num_samples = 100000
-    rates = Variable(torch.Tensor([rate]).expand(num_samples, 1), requires_grad=True)
-    factors = Variable(torch.Tensor([factor]).expand(num_samples, 1), requires_grad=True)
-    dist1 = Exponential(rates)
-    dist2 = RejectionExponential(rates, factors)
-    x1 = dist1.sample()
-    x2 = dist2.sample()
-    assert_equal(x1.mean(), x2.mean(), prec=0.02, msg='bug in .sample()')
-    assert_equal(x1.std(), x2.std(), prec=0.02, msg='bug in .sample()')
-    assert_equal(dist1.batch_log_pdf(x1), dist2.batch_log_pdf(x1), msg='bug in .batch_log_pdf()')
+    alphas = Variable(torch.Tensor([alpha]).expand(num_samples, 1), requires_grad=True)
+    betas = Variable(torch.ones(num_samples, 1))
+
+    model = Gamma(Variable(torch.ones(num_samples, 1)), betas)
+    guide1 = Gamma(alphas, betas)
+    guide2 = RejectionStandardGamma(alphas)  # implemented using Rejector
+
+    grads = []
+    for guide in [guide1, guide2]:
+        grads.append(compute_elbo_grad(model, guide, [alphas])[0].data)
+    expected, actual = grads
+    assert_equal(actual.mean(), expected.mean(), prec=0.01, msg='bad grad for alpha')
